+++ components/collection/drop/HolderOfGenerative.vue
@@ -40,8 +40,8 @@ import { fetchNft } from '@/components/items/ItemsGrid/useNftActions'
 import holderOfCollectionById from '@/queries/subsquid/general/holderOfCollectionById.graphql'
 import unlockableCollectionById from '@/queries/subsquid/general/unlockableCollectionById.graphql'
 import Loader from '@/components/shared/Loader.vue'
-import useGenerativeDropMint from './composables/useGenerativeDropMint'
-import useGenerativeDropDetails from './composables/useGenerativeDropDetails'
+import useGenerativeDropMint from '@/composables/drop/useGenerativeDropMint'
+import useGenerativeDropDetails from '@/composables/drop/useGenerativeDropDetails'
 
 const holderOfCollectionId = '50' // ChaosFlakes | todo: mock for testing, should be fetched from backend
