+++ composables/drop/useGenerativeDropDetails.ts
